test: split codex background refresh coalescing coverage
